+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.0.10] - 2023-04-07
+
+### Fixed
+
+- `VvFormWrapper` default slot `formData` and `errors` properties;
+- `defaultObjectBySchema` support for nested `ZodEffects`.
+
 ## [0.0.9] - 2023-03-23
 
 ### Added